finished cards

+++ lib/feature/home/models/game_response.dart
@@ -1,53 +1,54 @@
 import 'dart:convert';
-
 import 'package:team_play/feature/shared/models/create_by.dart';
 import 'package:team_play/feature/shared/models/location.dart';
 
-List<GameResponse> gameResponseFromJson(String str) => List<GameResponse>.from(json.decode(str).map((x) => GameResponse.fromJson(x)));
+List<GameResponse> gameResponseFromJson(String str) => List<GameResponse>.from(
+    json.decode(str).map((x) => GameResponse.fromJson(x)));
 
-String gameResponseToJson(List<GameResponse> data) => json.encode(List<dynamic>.from(data.map((x) => x.toJson())));
+String gameResponseToJson(List<GameResponse> data) =>
+    json.encode(List<dynamic>.from(data.map((x) => x.toJson())));
 
 class GameResponse {
-    final Location location;
-    final String id;
-    final String positionNeeded;
-    final CreatedBy createdBy;
-    final DateTime matchDate;
-    final String matchTime;
-    final int fieldRentalPayment;
-    final String? address;
-    final String description;
-    final DateTime createdAt;
-    final DateTime updatedAt;
-    final int v;
-    final String? playerInterested;
-    final String title;
+  final Location location;
+  final String id;
+  final String positionNeeded;
+  final CreatedBy createdBy;
+  final DateTime matchDate;
+  final String matchTime;
+  final double fieldRentalPayment;
+  final String? address;
+  final String description;
+  final DateTime createdAt;
+  final DateTime updatedAt;
+  final int v;
+  final String? playerInterested;
+  final String title;
 
-    GameResponse({
-        required this.location,
-        required this.id,
-        required this.positionNeeded,
-        required this.createdBy,
-        required this.matchDate,
-        required this.matchTime,
-        required this.fieldRentalPayment,
-        this.address,
-        required this.description,
-        required this.createdAt,
-        required this.updatedAt,
-        required this.v,
-        this.playerInterested,
-        required this.title,
-    });
+  GameResponse({
+    required this.location,
+    required this.id,
+    required this.positionNeeded,
+    required this.createdBy,
+    required this.matchDate,
+    required this.matchTime,
+    required this.fieldRentalPayment,
+    this.address,
+    required this.description,
+    required this.createdAt,
+    required this.updatedAt,
+    required this.v,
+    this.playerInterested,
+    required this.title,
+  });
 
-    factory GameResponse.fromJson(Map<String, dynamic> json) => GameResponse(
+  factory GameResponse.fromJson(Map<String, dynamic> json) => GameResponse(
         location: Location.fromJson(json["location"]),
         id: json["_id"],
         positionNeeded: json["position_needed"],
         createdBy: CreatedBy.fromJson(json["created_by"]),
         matchDate: DateTime.parse(json["match_date"]),
         matchTime: json["match_time"],
-        fieldRentalPayment: json["field_rental_payment"],
+        fieldRentalPayment: json["field_rental_payment"].toDouble(),
         address: json["address"],
         description: json["description"],
         createdAt: DateTime.parse(json["created_at"]),
@@ -55,9 +56,9 @@ class GameResponse {
         v: json["__v"],
         playerInterested: json["player_interested"],
         title: json["title"],
-    );
+      );
 
-    Map<String, dynamic> toJson() => {
+  Map<String, dynamic> toJson() => {
         "location": location.toJson(),
         "_id": id,
         "position_needed": positionNeeded,
@@ -72,5 +73,5 @@ class GameResponse {
         "__v": v,
         "player_interested": playerInterested,
         "title": title,
-    };
+      };
 }
